+++ contracts/interfaces/ITimeHelpers.sol
@@ -27,10 +27,10 @@ pragma solidity 0.6.10;
  * contract.
  */
 interface ITimeHelpers {
-    function addDays(uint fromTimestamp, uint n) external pure returns (uint);
-    function addMonths(uint fromTimestamp, uint n) external pure returns (uint);
-    function addYears(uint fromTimestamp, uint n) external pure returns (uint);
-    function timestampToDay(uint timestamp) external view returns (uint);
-    function timestampToMonth(uint timestamp) external view returns (uint);
-    function timestampToYear(uint timestamp) external view returns (uint);
+    function addDays(uint256 fromTimestamp, uint256 n) external pure returns (uint);
+    function addMonths(uint256 fromTimestamp, uint256 n) external pure returns (uint);
+    function addYears(uint256 fromTimestamp, uint256 n) external pure returns (uint);
+    function timestampToDay(uint256 timestamp) external view returns (uint);
+    function timestampToMonth(uint256 timestamp) external view returns (uint);
+    function timestampToYear(uint256 timestamp) external view returns (uint);
 }
